requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pgvector 
google-generativeai
streamlit
requests